Check if build_bin dir and ld symlink exist before trying to make them
If a previous build failed then the directory and symlink may already exist in
"$BUILD_DIR". If we try to make them again then we fail _differently_ on
subsequent build attempts, which is confusing.

+++ build.sh
@@ -42,7 +42,6 @@ if [ -f "${POSTGRES_PARSER_A}" ] && [ -d "${INSTALL_DIR}" ]; then
 fi
 
 if [ ! -f "${POSTGRES_LL}" ] ; then
-  mkdir -p "${BUILD_DIR}" || exit 1
   cd "${BUILD_DIR}" || exit 1
 
   # download/untar Postgres
@@ -58,8 +57,12 @@ if [ ! -f "${POSTGRES_LL}" ] ; then
   # configure, build, and (locally) install Postgres
   if [ "x${UNAME}" == "xLinux" ] ; then
     # linux needs to use the "gold" linker
-    mkdir build_bin || exit 1
-    ln -s /usr/bin/ld.gold build_bin/ld || exit 1
+    if [ ! -d build_bin ] ; then
+      mkdir build_bin || exit 1
+    fi
+    if [ ! -l build_bin/ld ] ; then
+      ln -s /usr/bin/ld.gold build_bin/ld || exit 1
+    fi
     CFLAGS="-B${PWD}/build_bin"
   fi
   AR="llvm-ar" CC="clang" CFLAGS="${CFLAGS} -flto" ./configure --without-readline --without-zlib --prefix="${INSTALL_DIR}" || exit 1
